Harden image decoders against DoS and add the Pentesting suite
Every decoder capped each axis but never the pixel product, so a header of
a few dozen bytes forced gigabyte allocations before any pixel data was
read. PNG/JPEG/GIF/NETPBM/DDS now share WEBP's total-pixel cap, and
readAll() bounds what one animation may retain.

Also closes two non-allocation classes: PNG compressed text spun forever
on a zlib stream requesting a preset dictionary (the inflate loop never
checked needsDictionary(), so it never progressed and never hit its own
size cap), and progressive JPEG, SVG rasterization and WEBP metadata each
had superlinear paths reachable from a small file.

Malformed input no longer escapes as IllegalArgumentException, raw
EOFException or bare IOException: the codec contract is XCodecException
throughout, so a corrupt image stays distinguishable from a transport
failure.

Adds Pentesting, driving 65 crafted images under src/test/resources/pentest
(regenerated by MaliciousImages), each modelled on a real CVE from a
C/C++/C# decoder. The corpus is marked -text so EOL conversion cannot
rewrite the bytes under test.

Code execution, local file read, SSRF and session-token exposure were
audited across every decoder and found unreachable; SVG's XXE posture is
pinned by regression fixtures.

+++ src/main/java/org/watermedia/api/codecs/writers/BCWriter.java
@@ -24,7 +24,7 @@ import java.nio.ByteBuffer;
 public final class BCWriter extends ImageWriter {
 
     private final String version;
-    private final int frameBytes;
+    private final long frameBytes;
     private long[] delays = new long[16];
     private int frames;
     private boolean closed;
